requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
packaging==26.0
PyJWT==2.12.1
Werkzeug==3.1.6
